refactor: MainTweetModal
