+++ src/components/MainCreateTweet/MainCreateTweet.jsx
@@ -7,7 +7,7 @@ const MainCreateTweet = ({ avatarImg }) => {
     <div className={styles.tweetInput}>
       <div className={styles.inputContainer} >
         <div className={styles.inputAvatar} >
-          <img
+          <MainTweetModal element={<img
             src={
               avatarImg ||
               "https://cdn-icons-png.flaticon.com/512/149/149071.png"
@@ -15,15 +15,18 @@ const MainCreateTweet = ({ avatarImg }) => {
             alt="user-avatar"
             width={50}
             height={50}
-          />
+          />} />
+          
         </div>
-        <input
+        <MainTweetModal element={<input
           className={styles.inputContent}
           placeholder="有什麼新鮮事？"
-        ></input>
+        ></input>} />
       </div>
       <div className={styles.modalBtn}>
-        <MainTweetModal />
+        <MainTweetModal element={<button className="btn btn-primary text-white rounded-pill">
+          推文
+      </button>}/>
       </div>
     </div>
   );
